move config stuff into separate file
